moves db to common
